requirements.txt
# Production WSGI server (for persistent server deployments only)
gunicorn>=22.0.0,<23.0.0

# Core dependencies
flask>=3.0.0,<4.0.0
flask-cors>=4.0.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0
openai>=1.0.0,<2.0.0

# PDF processing
PyPDF2>=3.0.0,<4.0.0
pdf2image>=1.16.0,<2.0.0
pillow>=10.0.0,<11.0.0
pytesseract>=0.3.10,<1.0.0

# RAG (vector embeddings and retrieval)
numpy>=1.24.0,<2.0.0
